Properly parse unicode characters in strings
This fixes the parsing of unicode strings that are 1 or 6
hex characters long.

Extends the tests to support a third element in examples, which
would be the re-stringified example, since a unicode character might
be stringified in a different way than the original code, and still be
valid.

fixes webpack/css-loader#133
fixes #12